Add shop: controlling class

+++ src/com/designpattern/model/Product.java
@@ -2,9 +2,9 @@ package com.designpattern.model;
 
 public class Product {
     private final String name;
-    private final long price;
+    private final int price;
 
-    public Product(String name, long price) {
+    public Product(String name, int price) {
         this.name = name;
         this.price = price;
     }
@@ -17,7 +17,7 @@ public class Product {
         return name;
     }
 
-    public long getPrice() {
+    public int getPrice() {
         return price;
     }
